+++ en/tools/quicklook/QuickLook.Master.cs
@@ -29,7 +29,7 @@ namespace SkyServer.Tools.QuickLook
         public QuickLookQueries exploreQuery;
 
         //public long? id = null;
-        public long? specId = null;
+        public decimal? specId = null;
         public string apid;
         public string objId = null;
         public string specObjId = null;
